pyproject.toml
[project]
name = "newsletter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "bleach>=6.1.0",
    "mysql-connector-python>=9.0.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "neovim>=0.3.1",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.1",
    "pre-commit>=3.5.0",
    "pytest-cov>=5.0.0",
]
